third_party/WebKit/LayoutTests/media/track/track-webvtt-tc032-degenerate-cues.html

<!DOCTYPE html>
<title>Tests that a degenerate cues without separating blank lines are parsed correctly.</title>
<script src="track-helpers.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
check_cues_from_track("captions-webvtt/tc032-degenerate-cues.vtt", function(track) {
    var expected = [
        { startTime: 0, endTime: 1, text: "" },
        { startTime: 2, endTime: 3, text: "" },
        { startTime: 4, endTime: 5, text: "" }
    ];

    assert_cues_match(track.cues, expected);
});
</script>
